requirements.txt
# Global Empathy News - Dependencies

# RSS feed parsing
feedparser>=6.0.0

# Google Gemini API
google-generativeai>=0.3.0
